Neste steg knapp på stønadsvilkår skal ferdsigstille VILKÅR-steget, og vise redigerknapp når steget er låst
